move code in qgtpytree from logic to main file
improts

f
